requirements.txt
fake-useragent==0.1.11
requests==2.24.0
beautifulsoup4==4.9.3
selenium==3.141.0
lxml==4.6.1
html5lib==1.1
pymongo==3.11.0
pillow==8.0.1
PyExecJS==1.5.1

